# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    DeploymentRequestResponse:
      type: object
      properties:
        deploymentRequest:
          $ref: "#/components/schemas/PdDeploymentRequestDto"
        deployment:
          $ref: "#/components/schemas/PdDeploymentDto"
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    ListDeploymentRequestsResponse:
      type: object
      properties:
        deploymentRequests:
          type: array
          items:
            $ref: "#/components/schemas/PdDeploymentRequestDto"
...
    PdDeploymentDto:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        applicationName:
          type: string
        version:
          type: string
        commitSha:
          type: string
        runId:
          type: string
        status:
          $ref: "#/components/schemas/PdDeploymentStatus"
        containerName:
          type: string
        detail:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        finishedAt:
          $ref: "#/components/schemas/Instant"
    PdDeploymentRequestDto:
      type: object
      properties:
        id:
          type: string
        applicationName:
          type: string
        version:
          type: string
        environmentId:
          type: string
        packageName:
          type: string
        priority:
          type: string
        repoId:
          type: string
        projectId:
          type: string
        qualityGate:
          $ref: "#/components/schemas/PdQualityGate"
        gateDetail:
          type: string
        deploymentId:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        gateSettledAt:
          $ref: "#/components/schemas/Instant"
        deploymentStatus:
          $ref: "#/components/schemas/PdDeploymentStatus"
    PdDeploymentStatus:
      type: string
      enum:
      - QUEUED
      - STARTING
      - ACTIVE
      - IMAGE_MISSING
      - SPEC_UNREADABLE
      - FAILED
      - ROLLED_BACK
      - SUPERSEDED
      - GONE
      - SCALED_TO_ZERO
      - DECLARATION_REFUSED
      - DECOMMISSIONED
...
    PdQualityGate:
      type: string
      enum:
      - UNMET
      - MET
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /deployments/api/deployment-requests:
    get:
      summary: "Deployment requests, newest-first — by environment, by project, or\
        \ by (repository, version)"
      operationId: listDeploymentRequests
      parameters:
      - name: applicationName
        in: query
        schema:
          type: string
      - name: environmentId
        in: query
        schema:
          type: string
      - name: projectId
        in: query
        schema:
          type: string
      - name: repoId
        in: query
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: string
      responses:
        "200":
          description: The requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDeploymentRequestsResponse"
        "400":
          description: "No filter, or half of the repoId/version pair"
        "404":
          description: No such environment
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Deployment Request Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
        - qits:system
  /deployments/api/deployment-requests/{id}:
    get:
      summary: "One deployment request, with the deployment it produced"
      operationId: getDeploymentRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentRequestResponse"
        "404":
          description: No such deployment request
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Deployment Request Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
        - qits:system
...
info:
  title: qits-platform-deployments API
  version: 2026.1008.164327